Implementação inicial do front.

+++ src/main/java/com/grupopti/estoque/repositories/ProdutoRepository.java
@@ -7,7 +7,7 @@ import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProdutoRepository extends CrudRepository<Produto, Integer> {
+public interface ProdutoRepository extends CrudRepository<Produto, Long> {
 
     List<Produto> findByDataCadastroBetween(Instant dataInicio, Instant dataFim);
     Optional<Produto> findByNome(String nome);
